Moved losses to own file. added dcgan d/g. dcwgan works with mnist, nothing else tested.
